+++ sql-plugin/src/main/31xdb/scala/org/apache/spark/rapids/shims/RapidsErrorUtils.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims
+package org.apache.spark.sql.rapids.shims
 
 import org.apache.spark.sql.catalyst.trees.Origin
 import org.apache.spark.sql.types.DataType
