+++ sql-plugin/src/main/321db/scala/org/apache/spark/rapids/shims/RapidsErrorUtils.scala
